chore: remove old gonotes, gonotes2 is now gonotes
