--- examples/renames/notes/2021-12/2021-12-21-2000-05-example-in-december.md
@@ -1,9 +1,0 @@
----
-title: Example in December
-date: 2021-12-21 20:00:05
-tags: example, journal, special/something/great
----
-
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse porttitor rutrum lectus, nec accumsan turpis lobortis at. Praesent eu rhoncus nibh. Nam dictum sagittis nisi at viverra. Ut mattis, diam in facilisis dictum, ligula lectus dapibus orci, eget placerat nulla felis sed nibh. Curabitur sodales rutrum eros vitae tempor. Sed quis dictum odio. Pellentesque metus nisl, egestas id laoreet quis, pellentesque non elit. Maecenas condimentum, neque a facilisis luctus, arcu sapien convallis turpis, non posuere diam risus id tellus. Aenean in nibh odio. Phasellus gravida orci eu ultrices ullamcorper. Proin hendrerit lacus non tristique rhoncus. Ut eleifend maximus eros eget gravida. Maecenas non rutrum erat. Nullam dictum risus eros, sed fringilla tortor auctor sit amet. In hac habitasse platea dictumst.
-
-Quisque ut tempor est. Donec posuere lectus luctus, blandit turpis vel, gravida arcu. Nullam non maximus sem. Nulla ultrices tortor id ligula varius, id sollicitudin erat convallis. Duis consequat justo sed lorem iaculis interdum. Praesent sit amet pretium augue, vel molestie metus. Sed sed ligula et diam feugiat mattis non vitae erat. Quisque urna dui, lacinia et eleifend id, congue pellentesque lacus. Cras mi eros, aliquet vitae nisi et, ornare volutpat lectus. Mauris quis lacus blandit, dictum turpis eu, accumsan neque. Donec iaculis in massa sit amet efficitur.
