--- examples/renames/notes/2021-12/2021-12-24-2231-48-ky.md
@@ -1,9 +1,0 @@
----
-title: Ky changed title and date
-date: 2021-10-24 22:31:48
-tags: bookmark/npm/request
-href: https://www.npmjs.com/package/ky
----
-
-Nog niet gebruikt, maar een axios alternatief dat er goed uit ziet!
-..
