--- examples/renames/notes/2022-01/2022-01-06-1705-52-multiple-meta-dilims.md
@@ -1,12 +1,0 @@
----
-title: Ky
-date: 2021-12-24 22:31:48
----
-
-Using meta dilims in content that should be ignored
-
----
-
-...
-
----
