--- examples/renames/tags/bookmark/npm/request/2021-12-24-2231-48-ky.md
@@ -1,1 +1,0 @@
-/Users/robotx/g/gonotes/examples/renames/notes/2021-12/2021-12-24-2231-48-ky.md
