--- examples/renames/tags/example/2021-12-21-2000-05-example-in-december.md
@@ -1,1 +1,0 @@
-/Users/robotx/g/gonotes/examples/renames/notes/2021-12/2021-12-21-2000-05-example-in-december.md
