--- examples/renames/tags/journal/2022-01-04-2155-26-example-one.md
@@ -1,1 +1,0 @@
-/Users/robotx/g/gonotes/examples/renames/notes/2022-01/2022-01-04-2155-26-example-one.md
