--- examples/synced/notes/2022-01/2022-01-06-1705-52-multiple-meta-dilims.md
@@ -1,12 +1,0 @@
----
-title: Multiple meta dilims
-date: 2022-01-06 17:05:52
----
-
-Using meta dilims in content that should be ignored
-
----
-
-...
-
----
