+++ go.mod
@@ -3,16 +3,6 @@ module github.com/marcelbeumer/gonotes
 go 1.25
 
 require (
-	github.com/bmatcuk/doublestar/v4 v4.4.0
-	github.com/spf13/cobra v1.6.1
-	github.com/stretchr/testify v1.8.1
+	github.com/google/go-cmp v0.7.0
 	gopkg.in/yaml.v3 v3.0.1
-)
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/google/go-cmp v0.7.0 // indirect
-	github.com/inconshreveable/mousetrap v1.0.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/spf13/pflag v1.0.5 // indirect
 )
