--- internal/cmd/error.go
@@ -1,11 +1,0 @@
-package cmd
-
-import (
-	"fmt"
-	"os"
-)
-
-func handleError(err error) {
-	fmt.Fprintln(os.Stderr, err.Error())
-	os.Exit(1)
-}
