--- internal/cmd/root.go
@@ -1,25 +1,0 @@
-package cmd
-
-import (
-	"os"
-
-	"github.com/spf13/cobra"
-)
-
-var verbose bool
-
-var rootCmd = &cobra.Command{
-	Use:   "gonotes",
-	Short: "Personal note taking program",
-}
-
-func Execute() {
-	err := rootCmd.Execute()
-	if err != nil {
-		os.Exit(1)
-	}
-}
-
-func init() {
-	rootCmd.PersistentFlags().BoolVarP(&verbose, "verbose", "v", false, "verbose output")
-}
